Revert "feat(ts) migrate `authenticateAndUpgradeRole` to TS"
This reverts commit 0e31a9a9cfcb5b48ae566cd4353bb2d077f7ff83.
